Warning compilation
Former-commit-id: b15a3357efb1288b591fb9520bdbeec4a04b5efd

+++ libraries/reconstruction/image_rotational_pca.cpp
@@ -140,8 +140,8 @@ void ProgImageRotationalPCA::produceSideInfo()
     Nimg = MDin.size();
     size_t Ydim, Zdim, Ndim;
     getImageSize(MDin, Xdim, Ydim, Zdim, Ndim);
-    Nangles = floor(360.0 / psi_step);
-    Nshifts = (2 * max_shift_change + 1) / shift_step;
+    Nangles = (int)floor(360.0 / psi_step);
+    Nshifts = (int)((2 * max_shift_change + 1) / shift_step);
     Nshifts *= Nshifts;
 
     // Construct mask
@@ -312,7 +312,7 @@ void threadApplyT(ThreadArgument &thArg)
                   (*(ptrWnode+6)) +=pixval*(*(ptrHblock+6));
                   (*(ptrWnode+7)) +=pixval*(*(ptrHblock+7));
                 }
-                for (int j=jmax; j<MAT_XSIZE(Wnode); ++j, ptrHblock+=1, ptrWnode+=1)
+                for (size_t j=jmax; j<MAT_XSIZE(Wnode); ++j, ptrHblock+=1, ptrWnode+=1)
                 (*(ptrWnode )) +=pixval*(*ptrHblock );
                 ++i;
               }
@@ -403,7 +403,7 @@ void threadApplyTt(ThreadArgument &thArg)
               applyGeometry(1,Iaux,mI,A,IS_INV,true);
 
               // Update Hblock
-              for (int j=0; j<MAT_XSIZE(Hblock); j++)
+              for (size_t j=0; j<MAT_XSIZE(Hblock); j++)
               {
                 double dotproduct=0;
                 const double *ptrIaux=MULTIDIM_ARRAY(Iaux);
@@ -472,7 +472,7 @@ int ProgImageRotationalPCA::QR()
   Matrix1D<double> qj1, qj2;
   int iBlockMax=MAT_XSIZE(F)/4;
 
-  for (int j1=0; j1<MAT_YSIZE(F); j1++)
+  for (size_t j1=0; j1<MAT_YSIZE(F); j1++)
   {
     F.getRow(j1,qj1);
     // Project twice in the already established subspace
@@ -497,7 +497,7 @@ int ProgImageRotationalPCA::QR()
           (*ptr1++)-=s12*(*ptr2++);
           (*ptr1++)-=s12*(*ptr2++);
         }
-        for (int i=iBlockMax*4; i<MAT_XSIZE(F); ++i)
+        for (size_t i=iBlockMax*4; i<MAT_XSIZE(F); ++i)
         (*ptr1++)-=s12*(*ptr2++);
       }
     }
